Extracted sub commands into their own classes
